requirements.txt

openai
python-dotenv
embedchain
streamlit
scrapegraphai
playwright
mlflow
pandas